alert add

+++ src/pages/Dashboard/ManageEvents/ManageEvents.js
@@ -15,7 +15,9 @@
     },[fresh])
 
     const handleDelete = (id)=>{
-        fetch(`http://localhost:5000/deleteEvent/${id}`, {
+        const proceed = window.confirm("Are you sure? delete for this Event")
+        if(proceed){
+            fetch(`http://localhost:5000/deleteEvent/${id}`, {
             method: "DELETE",
             headers: {
                 "content-type": 'application/json'
@@ -24,7 +26,11 @@
         .then(res=>res.json())
         .then(data=>{
             setFresh(data)
+            
         })
+
+        }
+        
 
     }
     return (
